move common properties _active/_components/_prefab/_persistNode to base node
